Add starter `WebSocket` client side routes
