+++ requirements.txt
@@ -5,6 +5,7 @@ astroid==2.11.2
 async-timeout==4.0.2
 asyncio==3.4.3
 attrs==21.4.0
+beautifulsoup4==4.11.1
 bidict==0.22.0
 certifi==2021.10.8
 charset-normalizer==2.0.12
@@ -21,6 +22,8 @@ Flask-SQLAlchemy==2.5.1
 flask-talisman==1.0.0
 fonttools==4.31.2
 frozenlist==1.3.0
+future==0.18.2
+google==3.0.0
 gunicorn==20.1.0
 idna==3.3
 iniconfig==1.1.1
@@ -34,7 +37,7 @@ MarkupSafe==2.1.1
 matplotlib==3.5.1
 mccabe==0.7.0
 multidict==6.0.2
-numpy==1.22.2
+numpy==1.22.3
 openpyxl==3.0.9
 packaging==21.3
 Pillow==9.1.0
@@ -44,7 +47,6 @@ psycopg==3.0.11
 psycopg2==2.9.3
 py==1.11.0
 pylint==2.13.5
-pymemcache==3.5.1
 pyparsing==3.0.7
 pytest==7.1.1
 python-binary-memcached==0.31.1
@@ -57,10 +59,13 @@ requests==2.27.1
 sgp4==2.20
 six==1.16.0
 skyfield==1.42
+soupsieve==2.3.2.post1
 SQLAlchemy==1.4.35
+toml==0.10.2
 tomli==2.0.1
 uhashring==2.1
 urllib3==1.26.8
+websocket-client==1.3.2
 websockets==10.3
 Werkzeug==2.1.1
 wrapt==1.14.0
